+++ src/hooks/useClearLocalStorage.ts
@@ -1,0 +1,9 @@
+import { useEffect } from "react";
+
+const useClearLocalStorage = () => {
+  useEffect(() => {
+    localStorage.clear();
+  }, []);
+};
+
+export { useClearLocalStorage };
