+++ src/pages/selectBusinessUnits/errors/ErrorNotBusinessUnit/index.tsx
@@ -1,14 +1,17 @@
-import { useEffect } from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 import { ErrorPage } from "@components/layout/ErrorPage";
 import clientNotFound from "@assets/images/Expired.png";
 import { enviroment } from "@config/environment";
+import { useClearLocalStorage } from "@hooks/useClearLocalStorage";
 
 function ErrorNotBusinessUnit() {
   const { logout } = useAuth0();
-  useEffect(() => {
-    localStorage.clear();
-  }, []);
+
+  useClearLocalStorage();
+
+  const handlelogout = () => {
+    logout({ logoutParams: { returnTo: enviroment.REDIRECT_URI } });
+  };
 
   return (
     <ErrorPage
@@ -16,9 +19,7 @@ function ErrorNotBusinessUnit() {
       imageAlt="Unidad de negocio no encontrada"
       heading="No hay resultados..."
       description="Su usuario no tiene unidades de negocio relacionados, consulte con su administrador."
-      onClick={() => {
-        logout({ logoutParams: { returnTo: enviroment.REDIRECT_URI } });
-      }}
+      onClick={handlelogout}
     />
   );
 }
