+++ src/pages/selectBusinessUnits/outlets/BusinessUnit/index.tsx
@@ -1,9 +1,6 @@
-import React, { useState, useContext } from "react";
-import { useNavigate } from "react-router-dom";
-import { AppContext } from "@context/AppContext";
 import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortalBusiness.types";
+import { useBusinessUnits } from "@hooks/selectBusinessUnits/useBusinessUnits";
 import { BusinessUnitsUI } from "./interface";
-import { IBusinessUnitstate } from "./types";
 
 interface IBusinessUnits {
   businessUnits: IBusinessUnitsPortalStaff[];
@@ -11,57 +8,15 @@ interface IBusinessUnits {
 
 function BusinessUnits(props: IBusinessUnits) {
   const { businessUnits } = props;
-  const { setBusinessUnitSigla } = useContext(AppContext);
-  const [search, setSearch] = useState("");
-  const [selectedBusinessUnit, setSelectedBusinessUnit] =
-    useState<IBusinessUnitsPortalStaff>();
-  const [businessUnitLocal, setBusinessUnitLocal] =
-    useState<IBusinessUnitstate>({
-      ref: null,
-      value: true,
-    });
-  const navigate = useNavigate();
 
-  const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (businessUnitLocal.ref) {
-      businessUnitLocal.ref.checked = false;
-    }
-    setBusinessUnitLocal({ ref: null, value: true });
-    setSearch(event.target.value);
-  };
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setBusinessUnitLocal({ ref: event.target, value: false });
-    const selectOption = businessUnits.find(
-      (businessUnit0) => businessUnit0.abbreviatedName === event.target.value,
-    );
-    setSelectedBusinessUnit(selectOption);
-  };
-
-  const handleSubmit = () => {
-    if (selectedBusinessUnit) {
-      const selectJSON = JSON.stringify(selectedBusinessUnit);
-      setBusinessUnitSigla(selectJSON);
-    }
-    navigate("/selectBusinessUnit/loading-app");
-  };
-
-  function filterBusinessUnits(
-    businessUnits: IBusinessUnitsPortalStaff[],
-    search: string,
-  ) {
-    const searchTerm = search?.toUpperCase();
-
-    return businessUnits.filter((unit) => {
-      const businessUnitName = unit?.abbreviatedName?.toUpperCase() || "";
-      const businessUnitSigla = unit?.publicCode?.toUpperCase() || "";
-
-      return (
-        businessUnitName.includes(searchTerm) ||
-        businessUnitSigla.includes(searchTerm)
-      );
-    });
-  }
+  const {
+    search,
+    businessUnitLocal,
+    handleSearchChange,
+    handleChange,
+    handleSubmit,
+    filterBusinessUnits,
+  } = useBusinessUnits(businessUnits);
 
   return (
     <BusinessUnitsUI
