+++ src/pages/selectBusinessUnits/outlets/CheckingCredentials/index.tsx
@@ -1,8 +1,6 @@
-import { useNavigate } from "react-router-dom";
-import { useCallback, useContext, useEffect } from "react";
-
-import { AppContext } from "@context/AppContext";
+import { useEffect } from "react";
 import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortalBusiness.types";
+import { useCheckCredentials } from "@hooks/selectBusinessUnits/useCredentials";
 import { CheckingCredentialsUI } from "./interface";
 
 function CheckingCredentials({
@@ -10,40 +8,10 @@ function CheckingCredentials({
 }: {
   businessUnits: IBusinessUnitsPortalStaff[];
 }) {
-  const navigate = useNavigate();
-  const { appData, setBusinessUnitSigla } = useContext(AppContext);
-
-  const selectedBusinessUnit = () => {
-    const selected = businessUnits[0];
-    setBusinessUnitSigla(JSON.stringify(selected));
-    navigate("/selectBusinessUnit/loading-app");
-  };
-
-  const checkCredentials = useCallback(() => {
-    try {
-      if (!appData) {
-        return;
-      }
-
-      if (appData) {
-        if (!businessUnits || businessUnits.length === 0) {
-          navigate("/selectBusinessUnit/error/not-related-businessUnits");
-        } else if (businessUnits.length === 1) {
-          selectedBusinessUnit();
-        } else {
-          navigate(`/selectBusinessUnit/businessUnits`);
-        }
-      } else {
-        navigate("/selectBusinessUnit/error/not-available");
-      }
-    } catch (error) {
-      console.info(error);
-      navigate("/selectBusinessUnit/error/not-available");
-    }
-  }, [appData, navigate, businessUnits]);
+  const checkCredentials = useCheckCredentials(businessUnits);
 
   useEffect(() => {
-    const timer = setTimeout(checkCredentials, 2000);
+    const timer = setTimeout(() => checkCredentials(), 2000);
     return () => clearTimeout(timer);
   }, [checkCredentials]);
